requirements.txt
httpx>=0.27
apscheduler>=3.10
anthropic>=0.30
google-genai>=1.0
python-dotenv>=1.0
pydantic-settings>=2.3
twilio>=9.0
